Abstracted out behaviour for pipeline request factories and models.
