+++ app/pacbio/pacbio.rb
@@ -5,4 +5,11 @@ module Pacbio
   def self.table_name_prefix
     'pacbio_'
   end
+
+  def self.attributes
+    %i[
+      library_type estimate_of_gb_required number_of_smrt_cells cost_code
+      external_study_id
+    ]
+  end
 end
