+++ app/saphyr/saphyr.rb
@@ -5,4 +5,10 @@ module Saphyr
   def self.table_name_prefix
     'saphyr_'
   end
+
+  def self.attributes
+    [
+      :external_study_id
+    ]
+  end
 end
