+++ spec/models/saphyr/request_spec.rb
@@ -2,8 +2,6 @@ require "rails_helper"
 
 RSpec.describe Saphyr::Request, type: :model, saphyr: true do
 
-  it 'must have an external study id' do
-    expect(build(:saphyr_request, external_study_id: nil)).to_not be_valid
-  end
+  it_behaves_like 'requestor model'
 
 end
